requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astroid==4.0.4
cachetools==7.0.3
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
dill==0.4.1
fastapi==0.135.1
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==8.0.1
loguru==0.7.3
mccabe==0.7.0
packaging==26.0
platformdirs==4.9.2
pluggy==1.6.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pylint==4.0.5
pytest==9.0.2
pytest-cov==7.0.0
python-multipart==0.0.22
python-dotenv==1.2.2
starlette==0.52.1
tenacity==9.1.4
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
uvloop==0.22.1; sys_platform != "win32"
